# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Darul Abror backend API (public + admin).",
        "title": "Darul Abror API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/registrations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registrations (Admin)"
                ],
                "summary": "Admin list registrations",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "new",
                            "validate",
                            "process",
                            "done"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.RegistrationListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/registrations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registrations (Admin)"
                ],
                "summary": "Admin get registration by ID",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Registration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.SuccessResponse-darulabror_internal_dto_RegistrationDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registrations (Admin)"
                ],
                "summary": "Admin delete registration",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Registration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/registrations/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registrations (Admin)"
                ],
                "summary": "Admin update registration status",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Registration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.RegistrationStatusUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/registrations": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registrations (Public)"
                ],
                "summary": "Create registration",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/darulabror_internal_dto.RegistrationDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "darulabror_internal_dto.RegistrationDTO": {
            "type": "object",
            "required": [
                "address",
                "date_of_birth",
                "date_of_birth_father",
                "date_of_birth_mother",
                "email",
                "father_name",
                "father_occupation",
                "full_name",
                "gender",
                "mother_name",
                "mother_occupation",
                "nisn",
                "origin_school",
                "phone",
                "phone_father",
                "phone_mother",
                "place_of_birth",
                "student_type"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "date_of_birth_father": {
                    "type": "string"
                },
                "date_of_birth_mother": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "father_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "father_occupation": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "gender": {
                    "enum": [
                        "male",
                        "female"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/darulabror_internal_models.Gender"
                        }
                    ]
                },
                "id": {
                    "type": "integer"
                },
                "mother_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "mother_occupation": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "nisn": {
                    "type": "string"
                },
                "origin_school": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "phone": {
                    "type": "string",
                    "maxLength": 13,
                    "minLength": 10
                },
                "phone_father": {
                    "type": "string",
                    "maxLength": 13,
                    "minLength": 10
                },
                "phone_mother": {
                    "type": "string",
                    "maxLength": 13,
                    "minLength": 10
                },
                "place_of_birth": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "status": {
                    "$ref": "#/definitions/darulabror_internal_models.RegistrationStatus"
                },
                "student_type": {
                    "enum": [
                        "new",
                        "transfer"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/darulabror_internal_models.StudentType"
                        }
                    ]
                }
            }
        },
        "darulabror_internal_models.Gender": {
            "type": "string",
            "enum": [
                "male",
                "female"
            ],
            "x-enum-varnames": [
                "Male",
                "Female"
            ]
        },
        "darulabror_internal_models.RegistrationStatus": {
            "type": "string",
            "enum": [
                "new",
                "validate",
                "process",
                "done"
            ],
            "x-enum-varnames": [
                "RegistrationStatusNew",
                "RegistrationStatusValidate",
                "RegistrationStatusProcess",
                "RegistrationStatusDone"
            ]
        },
        "darulabror_internal_models.StudentType": {
            "type": "string",
            "enum": [
                "new",
                "transfer"
            ],
            "x-enum-varnames": [
                "StudentNew",
                "StudentTransfer"
            ]
        },
...
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "something went wrong"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                }
            }
        },
...
        "internal_handler.ListResponseData-darulabror_internal_dto_RegistrationDTO": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/darulabror_internal_dto.RegistrationDTO"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/internal_handler.PaginationMeta"
                }
            }
        },
...
        "internal_handler.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
        "internal_handler.RegistrationListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_handler.ListResponseData-darulabror_internal_dto_RegistrationDTO"
                },
                "message": {
                    "type": "string",
                    "example": "OK"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "internal_handler.RegistrationStatusUpdateRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "new",
                        "validate",
                        "process",
                        "done"
                    ],
                    "example": "validate"
                }
            }
        },
...
        "internal_handler.SuccessResponse-darulabror_internal_dto_RegistrationDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/darulabror_internal_dto.RegistrationDTO"
                },
                "message": {
                    "type": "string",
                    "example": "OK"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}